Added 2 Documents and 1 Powerpoint
System Design document. Is graphs and diagrams I created for every part of scrum.
This includes:
2 UML activity Diagram
1 UML sequence Diagram
1 UML Class Diagram
1 UML Case Diagram

I also made a power point that's suppose to be low level for Product owners to understand showing the pros and cons for agile.

Speaker notes are to help you give the presentation. It's a more detailed version of the speaker notes on the slides themselves.